Add a Drawing Manager component
This is a drawing manager component that brings up a UI for drawing
shapes on your map. If you leave it empty, it will bring up the default
DrawingManager toolbar, but you can also add your own toolbar into the
default slot. In this case, slot scoped methods are provided to change
the drawing mode, and delete selected shapes.

The mandatory shapes prop will contain all the drawn shapes. If this has
existing shape data in it, they will be drawn onto the map when the
editor is displayed.

Please test the examples first!